Major cleanup of credit dialogs hierarchy, in progress
