+++ ui/app/components/credits-worksheet.hbs
@@ -11,7 +11,7 @@
       type="button"
       class="button"
       data-test-combine-credits
-      {{action "combineCredits"}}
+      onclick={{fn this.showCombine}}
       disabled={{combineLinkDisabled}}
     >
       <FaIcon @icon="object-group" class="no-underline mr-2" />
@@ -24,16 +24,37 @@
         <th></th>
         <th>Credit</th>
         <th>Term</th>
-        <th>Credits</th>
-        <th>Completed</th>
-        <th>Approved</th>
+        <th class="text-right">Credits</th>
+        <th class="text-right">Completed</th>
+        <th class="text-right">Approved</th>
         <th>Actions</th>
       </tr>
     </thead>
     <tbody>
       {{#each viewModels as |viewModel|}}
-        <CreditsWorksheet::Item @creditAssignment={{viewModel.creditAssignment}} @splitCredit={{action "splitCredit"}} @selectCredit={{action "selectCredit"}} @approveCredit={{action "approveCredit"}} />
+        <CreditsWorksheet::Item
+          @creditAssignment={{viewModel.creditAssignment}}
+          @splitCredit={{action "splitCredit"}}
+          @selectCredit={{action "selectCredit"}}
+          @approveCredit={{action "approveCredit"}}
+        />
       {{/each}}
     </tbody>
   </Table>
 {{/check-empty}}
+
+{{#if showCombineDialog}}
+  <ModalDialog
+    @onClose={{action "hideCombine"}}
+    @translucentOverlay={{true}}
+  >
+    <CreditAssignments::CombineDialog
+      @creditAssignments={{creditsToCombine}}
+      @model={{combineModel}}
+      @today={{today}}
+      @terms={{terms}}
+      @save={{action "combineCredits"}}
+      @close={{action (mut showCombineDialog) false}}
+    />
+  </ModalDialog>
+{{/if}}
