+++ ui/app/components/t-select.hbs
@@ -1,7 +1,6 @@
 <select
   ...attributes
   onchange={{fn this.onChange}}
-  onclick={{fn this.onClick}}
 >
   {{#if hasBlock}}
     {{yield
