+++ ui/app/templates/student-credits.hbs
@@ -7,12 +7,9 @@
 <h3>Available credits to approve</h3>
 
 <CreditsWorksheet
-  @student={{student}}
-  @creditAssignments={{creditAssignments}}
-  @selectedCredits={{selectedCredits}}
-  @showCombine={{action "showCombine"}}
-  @splitCredit={{action "splitCredit"}}
-  @approveCredit={{action "approveCredit"}}
+  @student={{this.student}}
+  @creditAssignments={{this.creditAssignments}}
+  @updateCreditAssignments={{fn this.updateCreditAssignments}}
 />
 
 <h3>Credits finalized and sent to district</h3>
@@ -21,16 +18,3 @@
   @student={{student}}
   @creditAssignments={{creditAssignments}}
 />
-
-{{#if showCombineDialog}}
-  <ModalDialog @onClose={{action "hideCombine"}} @translucentOverlay={{true}}>
-    <CreditAssignments::CombineDialog
-      @creditAssignments={{creditsToCombine}}
-      @model={{combineModel}}
-      @today={{today}}
-      @terms={{terms}}
-      @save={{action "combineCredits"}}
-      @close={{action (mut showCombineDialog) false}}
-    />
-  </ModalDialog>
-{{/if}}
